fix: Rename contacts to contact (singular)
