+++ packages/cozy-client/src/models/index.js
@@ -7,7 +7,7 @@ import * as account from './account'
 import * as note from './note'
 import * as permission from './permission'
 import * as utils from './utils'
-import * as contacts from './contacts'
+import * as contact from './contact'
 
 // For backward compatibility before 9.0.0
 const triggers = trigger
@@ -25,5 +25,5 @@ export {
   accounts,
   permission,
   utils,
-  contacts
+  contact
 }
